Added the exe, readme and cpp file
Added the Readme file which shows the user what to do on the github page.

Added the color_picker.exe which runs on windows machines aswell as linux and mac if they are running wine.

Added the color_picker.cpp which is the actual code for the program. This is what makes it run and linux and mac users can complie if required